ase/buffer: fix allocation on systems with default 1GB hugepage
When a MAP_HUGETLB mapping is requested without specifying a
desired huge page size, it will use the system's default huge
page size. This means that on a system with default 1GB huge
pages, a request for 2MB can be rounded up to 1GB. Explicitly
specify the size in all cases to prevent this from occurring.